Add the option to "integrate" a metric into a Prometheus counter
The metrics we receive from Cloudwatch are represented as gauges.
However, for some metrics such as the sum of requests to a cloudfront
distribution, it would be more useful to have the metrics as counters in
Prometheus. This allows for more natural integration with other tools
such as SLO management tools such as Pyrra. It should also get rid of
problems with matching the prometheus scrape interval with the YACE
scrape interval to avoid repeated data points. A repeated value for a
counter just means the value hasn't increased, and can be managed using
appropriate time ranges on "rate" or "increase" functions.